requirements.txt
pymilvus>=2.4.0
numpy>=1.21.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
